+++ tests/kubernetes_test.py
@@ -184,14 +184,14 @@ class TestChecks:
     async def test_check_version(self, config) -> None:
         await config.load_kubeconfig()
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id="check_version"))
+        results = await checks.run_all(matching=servo.checks.Filter(id="check_version"))
         assert results
         assert results[-1].success
 
     async def test_check_connectivity_success(self, config) -> None:
         await config.load_kubeconfig()
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id="check_connectivity"))
+        results = await checks.run_all(matching=servo.checks.Filter(id="check_connectivity"))
         assert len(results) == 1
         assert results[0].success
     
@@ -205,7 +205,7 @@ class TestChecks:
         KubernetesAsyncioConfiguration.set_default(loaded_config)
 
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id="check_connectivity"))
+        results = await checks.run_all(matching=servo.checks.Filter(id="check_connectivity"))
         assert len(results) == 1
         result = results[0]        
         assert not result.success
@@ -214,7 +214,7 @@ class TestChecks:
     async def test_check_permissions_success(self, config: KubernetesConfiguration) -> None:
         await config.load_kubeconfig()
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id=["check_permissions"]))
+        results = await checks.run_all(matching=servo.checks.Filter(id=["check_permissions"]))
         assert len(results)
         result = results[-1]
         assert result.id == "check_permissions"
@@ -226,7 +226,7 @@ class TestChecks:
     async def test_check_namespace_success(self, config) -> None:
         await config.load_kubeconfig()
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id=["check_namespace"]))
+        results = await checks.run_all(matching=servo.checks.Filter(id=["check_namespace"]))
         assert len(results)
         result = results[-1]
         assert result.id == "check_namespace"
@@ -236,7 +236,7 @@ class TestChecks:
         await config.load_kubeconfig()
         config.namespace = "INVALID"
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id=["check_namespace"]))
+        results = await checks.run_all(matching=servo.checks.Filter(id=["check_namespace"]))
         assert len(results)
         result = results[-1]
         assert result.id == "check_namespace"
@@ -256,7 +256,7 @@ class TestChecks:
         await config.load_kubeconfig()
         config.deployments[0].name = "INVALID"
         checks = kubernetes_connector.KubernetesChecks(config)
-        results = await checks.run_(servo.checks.Filter(id=["check_deployments_item_0"]))
+        results = await checks.run_all(matching=servo.checks.Filter(id=["check_deployments_item_0"]))
         assert len(results)
         result = results[-1]
         assert result.id == "check_deployments_item_0"